# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/channels/me": {
      "get": {
        "operationId": "ChannelsController_getMyChannel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get my channel",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "operationId": "ChannelsController_updateMyChannel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update my channel info (name, nickname, description)",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/videos": {
      "get": {
        "description": "Returns paginated public videos with optional search and category filter.",
        "operationId": "VideosController_listVideos",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedVideosDto"
                }
              }
            }
          }
        },
        "summary": "List public videos",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}": {
      "get": {
        "description": "Returns video details. Works for both public and unlisted videos.",
        "operationId": "VideosController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Get video by slug",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/videos/{id}": {
      "patch": {
        "description": "Updates title, description, or category. Owner only.",
        "operationId": "VideosController_updateVideo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update video info",
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "description": "Permanently deletes the video and its files from storage. Owner only.",
        "operationId": "VideosController_deleteVideo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete video",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/videos/channel/me": {
      "get": {
        "description": "Returns all videos for the authenticated user's channel (management panel).",
        "operationId": "VideosController_listMyVideos",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelVideosDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my channel videos",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/videos/{slug}/like": {
      "post": {
        "operationId": "VideoLikesController_like",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Like a video",
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "operationId": "VideoLikesController_removeLike",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove like/dislike from a video",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/videos/{slug}/comments": {
      "get": {
        "operationId": "CommentsController_listComments",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCommentsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "List comments for a video",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Post a comment on a video",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChannelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnail_key": {
            "type": "string",
            "nullable": true
          },
          "user_id": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "nickname",
          "user_id",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateChannelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Channel"
          },
          "nickname": {
            "type": "string",
            "example": "mychannel"
          },
          "description": {
            "type": "string",
            "example": "My channel description"
          }
        }
      },
...
      "ChannelSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "nickname",
          "name"
        ]
      },
      "CategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "VideoCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true
          },
          "duration_seconds": {
            "type": "number",
            "nullable": true
          },
          "view_count": {
            "type": "number"
          },
          "published_at": {
            "format": "date-time",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelSummaryDto"
          },
          "category": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorySummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "view_count",
          "published_at",
          "channel"
        ]
      },
      "PaginatedVideosDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoCardDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total_pages": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "total_pages"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "unlisted"
            ],
            "nullable": true
          },
          "duration_seconds": {
            "type": "number",
            "nullable": true
          },
          "view_count": {
            "type": "number"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelSummaryDto"
          },
          "category": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategorySummaryDto"
              }
            ]
          },
          "published_at": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "status",
          "view_count",
          "channel",
          "created_at"
        ]
      },
      "UpdateVideoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "My awesome video"
          },
          "description": {
            "type": "string",
            "example": "A description of the video"
          },
          "category_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ChannelVideosDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoResponseDto"
            }
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "LikeStatusDto": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "number"
          },
          "dislikes": {
            "type": "number"
          },
          "userLike": {
            "type": "string",
            "enum": [
              "like",
              "dislike"
            ],
            "nullable": true
          }
        },
        "required": [
          "likes",
          "dislikes"
        ]
      },
      "CommentAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/CommentAuthorDto"
          },
          "parent_id": {
            "type": "string",
            "nullable": true
          },
          "deleted": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "body",
          "author",
          "parent_id",
          "deleted",
          "created_at",
          "updated_at"
        ]
      },
      "PaginatedCommentsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            }
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "example": "Great video!"
          },
          "parent_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "body"
        ]
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}